+++ ex2_hal/adcs/equipment_handler/src/adcs_handler.c
@@ -708,7 +708,7 @@ ADCS_returnState ADCS_get_EDAC_err_count(uint16_t* single_sram,
  * @return
  * 		Success of function defined in adcs_types.h
  */
-ADCS_returnState ADCS_get_comms_stat(uint8_t telemetry[6], uint16_t* TC_num, uint16_t* TM_num,
+ADCS_returnState ADCS_get_comms_stat(uint16_t telemetry[6], uint16_t* TC_num, uint16_t* TM_num,
                                      uint8_t* flags_arr) {
   ADCS_returnState state;
   state = adcs_telemetry(COMMS_STAT_ID, telemetry, 6);
